+++ tests/lib/rules/multiline-object-properties.js
@@ -9,14 +9,8 @@ ruleTester.run('multiline-object-properties', rule, {
     'x={\na:1,\nb:2,\nc:3\n}',
   ],
   invalid: [{
-    code: 'x={a:1,b:2,c:3}',
-    errors: [1, 2, 3, 'and finally the end'].map(() => ({
-      message: 'Property in an object should be on a new line.',
-      type: 'ObjectExpression',
-    })),
-  }, {
-    code: 'x={a:\'looooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong\'}',
-    errors: [1, 'and finally the end'].map(() => ({
+    code: 'x={a:1,\nb:2,c:3}',
+    errors: [2, 3, 'and finally the end'].map(() => ({
       message: 'Property in an object should be on a new line.',
       type: 'ObjectExpression',
     })),